demos/demo2-mcp-tool-abuse/pyproject.toml
[project]
name = "demo2-mcp-tool-abuse"
version = "0.1.0"
description = "Demo: MCP Tool Description Poisoning Attack"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flock-core",
    "python-dotenv>=1.0.0",
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
